+++ tests/conftest.py
@@ -223,7 +223,7 @@ def fixture_mock_quum_api_handler(
     """A logged-in QuantinuumQAPI fixture.
     After using this fixture in a test, call:
         mock_quum_api_handler.delete_authentication()
-    To remove mock tokens from the keyring.
+    To remove mock tokens from memory.
     """
 
     username, pwd = mock_credentials
@@ -241,16 +241,13 @@ def fixture_mock_quum_api_handler(
     )
 
     cred_store = MemoryCredentialStorage()
-    cred_store._save_login_credential(
-        user_name=username,
-        password=pwd,
-    )
+    cred_store.save_user_name(username)
+    cred_store._password = pwd
 
     # Construct QuantinuumQAPI and login
     api_handler = QuantinuumAPI()
 
     # Add the credential storage seperately in line with fixture parameters
-    api_handler.config.username = username
     api_handler._cred_store = cred_store
     api_handler.login()